+++ ICD.Connect.UI/Attributes/AbstractUiAttribute.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Common.Utils.Attributes;
+
+namespace ICD.Connect.UI.Attributes
+{
+	public abstract class AbstractUiAttribute : AbstractIcdAttribute
+	{
+	}
+}