+++ ICD.Connect.UI/Attributes/ViewBindingAttribute.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using ICD.Connect.UI.Mvp.Views;
+
+namespace ICD.Connect.UI.Attributes
+{
+	public sealed class ViewBindingAttribute : AbstractUiBindingAttribute<IView>
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="interfaceBinding"></param>
+		public ViewBindingAttribute(Type interfaceBinding)
+			: base(interfaceBinding)
+		{
+		}
+	}
+}