+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added view/presenter bindings for automatically mapping interfaces to concretes
  - Added view/presenter interface binding cache
  - AbstractViewFactory provides a default method for instantiating a view by interface
+ - AbstractNavigationController provides a default method for instantiating a presenter by interface
 
 ## [5.0.0] - 2018-11-20
 ### Added